:hammer: refactored enrollment
